% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSTANCEobject.R
\name{creatSTANCEobject}
\alias{creatSTANCEobject}
\title{Create the STANCE object.}
\usage{
creatSTANCEobject(counts, pos, prop, covariates = NULL)
}
\arguments{
\item{counts}{original gene expression matrix of dimension n x G,
with genes in rows and spots in columns.}

\item{pos}{the n x 2 dimensional location coordinates matrix.
The rownames of pos matrix should match the colnames of counts matrix.}

\item{prop}{the cell type proportion of each spot (n x K matrix)
Rows represent spots and columns represent cell types.
The rownames of prop matrix should match the colnames of counts matrix.}

\item{covariates}{(default NULL) the covariate (if any) design matrix of dimension n x p,
modeling gene expression.
The rownames of covariates matrix should match the colnames of counts matrix.}
}
\value{
returns STANCE object.
}
\description{
Create the STANCE object.
}
