% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{data_preprocess}
\alias{data_preprocess}
\title{Preprocess data}
\usage{
data_preprocess(
  object,
  spot.threshold = 10,
  gene.threshold = 0.05,
  normalized = FALSE
)
}
\arguments{
\item{object}{STANCE object}

\item{spot.threshold}{(default 10) filter out the spots whose total expression count
lower than the threshold.}

\item{gene.threshold}{(default 0.05) filter out low-expressed genes
whose expression rate lower than the threshold.}

\item{normalized}{(default FALSE) "normalized = TRUE" indicates that the gene expression
input is normalized. Skip the normalization step.}
}
\value{
return STANCE object.
}
\description{
Preprocess the gene expression, location coordinate and cell type proportion matrices,
including normalization of count data (use NormalizeVST in SPARK package),
scaling of coordinates and quality control at both spot and gene levels.
}
