% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modelMatrices.R
\name{get_modelMatrices}
\alias{get_modelMatrices}
\title{Construct the model matrices}
\usage{
get_modelMatrices(object)
}
\arguments{
\item{object}{the STANCE object.}
}
\value{
return STANCE object.
}
\description{
Construct the important matrices, including design matrices, covariance matrices
and Sigma_k matrices for all the random effects, before fitting STANCE model.
}
