% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.R
\name{invert}
\alias{invert}
\title{Invert a matrix using RcppArmadillo}
\usage{
invert(mat)
}
\arguments{
\item{mat}{A numeric matrix to be inverted}
}
\value{
The inverted matrix
}
\description{
Invert a matrix using RcppArmadillo
}
\examples{
invert(matrix(c(1, 2, 3, 4), nrow = 2))
}
