% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mOB}
\alias{mOB}
\title{mOB}
\format{
An object of class \code{list} of length 3.
}
\usage{
mOB
}
\description{
Mouse olfactory bulb spatial transcriptomics dataset
}
\details{
A cleaned dataset containing the gene expression, spatial coordinates of spots,
and cell type compositions of mouse olfactory bulb replicate No.8.
Cell type compositions are based on cell type doconvolution through STdeconvolve.
}
\examples{
  str(mOB)
  head(mOB$counts)
  head(mOB$pos)
  head(mOB$prop)

}
\keyword{datasets}
