% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTest1.R
\name{runTest1}
\alias{runTest1}
\title{Perform STANCE overall tests for detecting SVGs and ctSVGs}
\usage{
runTest1(object, correction = FALSE, pv.adjust = "BY")
}
\arguments{
\item{object}{the STANCE object.}

\item{correction}{(default FALSE) if TRUE, perform a bias correction
for the approximation of the scaled chi-square distribution.}

\item{pv.adjust}{(default "BY") p-value adjustment method, a character string.}
}
\value{
the STANCE object with overall test results.
}
\description{
Perform STANCE overall tests for detecting SVGs and ctSVGs
}
