% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_class.R
\docType{class}
\name{STANCE-class}
\alias{STANCE-class}
\title{An S4 class to store key information.}
\description{
An S4 class to store key information.
}
\section{Slots}{

\describe{
\item{\code{gene_expression}}{The gene expression matrix with genes in rows
and spots in columns.}

\item{\code{location}}{The location coordinate matrix of spots.}

\item{\code{cell_type_compositions}}{The cell type composition matrix
to store cell type proportion at each spot.
Rows represent spots and columns represent cell types.}

\item{\code{covariates}}{The covariate design matrix modeling gene expression.}

\item{\code{cell_types}}{Cell types included in this experiment.}

\item{\code{kernel}}{The kernel matrix to capture spatial correlation.
The default is a Gaussian kernel matrix.}

\item{\code{bandwidth}}{The bandwidth parameter of Gaussian kernel.}

\item{\code{Sigma_k_matrices}}{A list for Sigma_k matrices.}

\item{\code{Test_1}}{The results of STANCE Test 1 (overall test for SVGs and ctSVGs),
containing the test statistic values, p values and possibly adjusted p values
for each gene.}

\item{\code{Test_2}}{The results of STANCE Test 2 (test for ctSVGs of each cell type of interest),
containing the test statistic values, p values and possibly adjusted p values
for each gene and each cell type of interest.}

\item{\code{VarComp_estimates}}{The ReML estimates of all the variance components for each gene.}

\item{\code{cell_type_top_genes}}{The top significant genes with highest proportion of variance
explained by the spatial effect corresponding to the cell type of interest.}

\item{\code{original_gene_expression}}{Used only after the normalization of expression count data
and the quality control steps,
so as to store the original gene expression count matrix.}

\item{\code{original_location}}{Used only after the scaling procedure of location coordinates
and the quality control steps,
so as to store the original spot location matrix.}

\item{\code{original_cell_type_compositions}}{Used only the quality control steps,
so as to store the original spot location matrix.}
}}

