% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varcomp_est.R
\name{varcomp_est}
\alias{varcomp_est}
\title{Estimate the variance components for STANCE model}
\usage{
varcomp_est(object, Genes_to_est = NULL, ncores = 1)
}
\arguments{
\item{object}{STANCE object}

\item{Genes_to_est}{(default NULL) a vector of character strings
claiming all the genes of interest to estimate.
If NULL, estimation of variance components will perform across all the genes.}

\item{ncores}{(default 1) an integer value of number of CPU cores to use
when running variance component estimation.}
}
\value{
return STANCE object.
}
\description{
Estimate all the variance components of the full STANCE model for each gene of interest.
}
