% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeGenePattern.R
\name{visualizeGenePattern}
\alias{visualizeGenePattern}
\title{Visualize the expression levels of a specified gene across spots within the tissue}
\usage{
visualizeGenePattern(
  object,
  gene_to_display,
  scaled_gene_expression = TRUE,
  point_size = 4,
  filling_colors = c("white", "darkblue"),
  title_size = 20,
  legend_title_size = 18,
  legend_text_size = 16
)
}
\arguments{
\item{object}{STANCE object}

\item{gene_to_display}{character string claiming the gene to display.}

\item{scaled_gene_expression}{(default TURE) if TRUE, the gene expression levels
will be scaled to be the values between 0 and 1.}

\item{point_size}{(default 0.05) the point size on the pattern.}

\item{filling_colors}{(default c("white", "blue")) a 2-dimensional character string
vector to specify the "low" and "high" inputs respectively for "scale_fill_gradient()"
of ggplot2, which is to create a diverging colour gradient.}

\item{title_size}{(default 20) font size of the figure title.}

\item{legend_title_size}{(default 18) font size of the legend title.}

\item{legend_text_size}{(default 16) font size of the figure texts.}
}
\value{
return STANCE object.
}
\description{
visualize the expression levels of a specified gene across spatial locations within the tissue, providing an intuitive representation of spatial gene expression patterns.
}
