% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CT_topGenes.R
\name{CT_topGenes}
\alias{CT_topGenes}
\title{Display top genes for the cell type of interest given}
\usage{
CT_topGenes(object, CT_of_interest, numTopGenes = 20)
}
\arguments{
\item{object}{STANCE object}

\item{CT_of_interest}{character string claiming cell type of interest to display.}

\item{numTopGenes}{(default 20) number of top genes to display.
If the number of significant genes is less than this given number,
then all significant genes will be displayed.}
}
\value{
return STANCE object.
}
\description{
For the cell type of interest given, list out genes significant in the corresponding
STANCE individual test with highest proportion of variance explained by this cell type.
\'cell_type_of_interest\' indicates the cell type of interest given.
\'top_genes\' lists out all the top genes.
\'stacked_bar_plot\' is a stacked bar plot for top genes
visualizing the proportion of variance explained by all the cell types and error.
}
