% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTest2.R
\name{runTest2}
\alias{runTest2}
\title{Perform STANCE individual tests for detecting ctSVGs}
\usage{
runTest2(
  object,
  Cell_types_to_test = NULL,
  Genes_to_test = NULL,
  correction = FALSE
)
}
\arguments{
\item{object}{STANCE object}

\item{Cell_types_to_test}{(default NULL) a vector of character strings
claiming all the cell types of interest to test.
If NULL, tests will perform across all the cell types.}

\item{Genes_to_test}{(default NULL) a vector of character strings
claiming all the genes of interest to test.
If NULL, tests will perform across all the genes.}

\item{correction}{(default FALSE) if TRUE, perform a bias correction
for the approximation of the scaled chi-square distribution.}
}
\value{
return STANCE object.
}
\description{
Perform STANCE individual tests to detect ctSVGs for each cell type of interest given.
}
